+++ src/store/middleware/logoutMiddleware.ts
@@ -5,11 +5,15 @@ import { localStorageKeys } from 'storage/local';
 import { accountInfoSelector, isLoggedInSelector } from 'store/selectors';
 import { initialState as initialAccountState } from 'store/slices/account/accountSlice';
 import { initialState as initialLoginInfoState } from 'store/slices/loginInfo/loginInfoSlice';
+import { initialState as initialToastState } from 'store/slices/toast/toastSlice';
+import { initialState as initialTrackedTransactionsState } from 'store/slices/trackedTransactions/trackedTransactionsSlice';
 import { StoreType } from '../store.types';
 
 export const resetStore = (store: WritableDraft<StoreType>) => {
   store.account = initialAccountState;
   store.loginInfo = initialLoginInfoState;
+  store.trackedTransactions = initialTrackedTransactionsState;
+  store.toasts = initialToastState;
 };
 
 export function getNewLoginExpiresTimestamp() {